Allow multiple classes in one phpunit test file.
PHPUnit can properly handle this case without crashing.
Paratest should as well.

Before, Paratest would assume that the first class in the for loop was a
TestCase instance, and sometimes fail.

In several files in a project I work on,
there are helper classes alongside the test file,
used only in that test file.

Update test cases to reflect the new behaviour, require them to be
subclasses of TestClass